LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v2/script-src/manifest-v2-script-src-hash-source-allowed.html

<!DOCTYPE html><!-- webkit-test-runner [ contentSecurityPolicyExtensionMode=v2 ] -->
<html>
    <head>
        <script>
            if (window.testRunner)
                testRunner.dumpAsText();
            alert('PASS (1/4)');
        </script>
        <meta http-equiv="Content-Security-Policy" content="script-src 'sha256-j+ToFnhur5mTSCwoyy4Fh29+1BNMVEZ0MBvG5db0dA4=' 'sha384-S/1WrU/8y14JzSAhjkxjd8sZLP10LDlHuH8Mi24OwMNJ6wSry7J8ln0KeaibHHBW' 'sha512-0tQJUSPHRGUYQIa2ByQWZSlUa5sI3e2tOEXF7sg5mjmsHk/EHDTTORIDyMPhDKgCmdYdb7TIrwhLrq2kEvfjHQ=='">
        <script>
            alert('PASS (2/4)');
        </script>
        <script>
            alert('PASS (3/4)');
        </script>
        <script>
            alert('PASS (4/4)');
        </script>
    </head>
    <body>
        <p>
            Tests CSP for manifest v2 allows hash sources.
        </p>
    </body>
</html>
