LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v2/script-src/manifest-v2-script-src-hash-source-allowed-expected.txt

ALERT: PASS (1/4)
ALERT: PASS (2/4)
ALERT: PASS (3/4)
ALERT: PASS (4/4)
Tests CSP for manifest v2 allows hash sources.
